+++ src/components/layouts/index.ts
@@ -1,5 +1,2 @@
 export { default as ArticleLayout } from "./ArticleLayout";
 export { default as DefaultLayout } from "./DefaultLayout";
-export { default as HomeLayout } from "./HomeLayout";
-export { RootLayout } from "./RootLayout";
-export { TwoColumnLayout } from "./TwoColumnLayout";
